pyproject.toml
[project]
name = "black-trivia-project"
version = "0.1.0"
description = "Black Card Django backend for Vercel"
requires-python = ">=3.12,<3.13"
dependencies = [
  "Django==5.1.4",
  "djangorestframework==3.15.2",
  "djangorestframework-simplejwt==5.3.1",
  "django-cors-headers==4.3.1",
  "google-auth>=2.35,<3",
  "psycopg2-binary==2.9.10",
  "gunicorn==21.2.0",
  "requests>=2.32,<3",
  "sqlparse==0.5.3",
  "asgiref==3.8.1",
  "tzdata==2023.3",
]

[tool.vercel.scripts]
build = "python build.py"
